Added StringBuilderBenchmark
A long time ago, I wrote the following article on `StringBuffer`
versus `StringBuilder`:

https://alblue.bandlem.com/2016/04/jmh-stringbuffer-stringbuilder.html

The results are still in line with what was written, though the numbers are a
little bit different than before. Some of the GC churn has decreased, but the
relative performances are still approximately the same.

Results are in the JMH-StringBuilderBenchmark.txt file.